refactor: librarify `lurk` (#21)
